Set a default value to `options.killTimeout`
This line is incorrect:

  if (typeof options.killTimeout !== 'number') { options.timeout = 5000; }

It checks to see if `options.killTimeout` is not a number (most likely it
will be undefined), and then sets the `options.timeout` variable. This means
that for the setTimeout will still always have an undefined timeout, this
results in the setTimeout being executed immediately. Rendering both the
"immediate" and the "setTimeout" features here useless.

The only way to actually use the killTimeout option is to explicitly
set the option currently.

Closes #80